% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gofobj.R
\name{plot.gofobj}
\alias{plot.gofobj}
\title{Plot GoF object}
\usage{
\method{plot}{gofobj}(x, parm = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{"gofobj"}}

\item{parm}{do all in one plots}

\item{...}{other}
}
\description{
Function to plot an object of class \code{'gofobj'}
}
\examples{
Y <- network(20, directed = FALSE)[,]

modLSM <- lsm(Y, D = 2) 
myGof <- goflsm(modLSM, Y = Y, doplot = FALSE)
plot(myGof)
}
