% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{PPIgen}
\alias{PPIgen}
\title{PPI genetic interactions}
\format{Binary adjacency matrix}
\description{
The dataset contains a network formed by genetic protein-protein
interactions (PPI) between 67 Saccharomyces cerevisiae proteins.
The network is formed of 294 links.
The data were downloaded from the Biological General Repository for
Interaction Datasets (BioGRID) database \url{http://thebiogrid.org/}
}
\details{
Binary adjacency matrix containing genetic interactions between 67 proteins.
}
\references{
Gollini, I., and Murphy, T. B. (2014), "Joint Modelling of Multiple Network Views", Journal of Computational and Graphical Statistics \url{http://arxiv.org/abs/1301.3759}.
}
\seealso{
\code{\link{PPIphy}}
}
\keyword{datasets}

