% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIphy.R
\docType{data}
\name{PPIphy}
\alias{PPIphy}
\title{PPI physical interactions}
\format{Binary adjacency matrix}
\description{
The dataset contains a network formed by physical protein-protein
interactions (PPI) between 67 Saccharomyces cerevisiae proteins.
The network is formed of 190 links.
The data were downloaded from the Biological General Repository for
Interaction Datasets (BioGRID) database \url{http://thebiogrid.org/}
}
\details{
Binary adjacency matrix containing physical interactions between 67 proteins.
}
\references{
Gollini, I., and Murphy, T. B. (2016), 'Joint Modelling of Multiple Network Views', Journal of Computational and Graphical Statistics, 25(1), 246-265 \url{http://arxiv.org/abs/1301.3759}.
}
\seealso{
\code{\link{PPIgen}}
}
\keyword{datasets}
