% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gofobj.R
\name{print.gofobj}
\alias{print.gofobj}
\title{Print GoF object}
\usage{
\method{print}{gofobj}(x, ...)
}
\arguments{
\item{x}{object of class \code{'gofobj'}}

\item{...}{other}
}
\description{
Function to print an object of class \code{'gofobj'}
}
\examples{
Y <- network(20, directed = FALSE)[,]

modLSM <- lsm(Y, D = 2) 
myGof <- goflsm(modLSM, Y = Y, doplot = FALSE)
print(myGof)
}
