% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotXtoY.R
\name{rotXtoY}
\alias{rotXtoY}
\title{Rotate X to match Y}
\usage{
rotXtoY(X, Y)
}
\arguments{
\item{X}{matrix to be rotated}

\item{Y}{objective matrix}
}
\value{
rotated object \code{Xrot}, and the rotation matrix \code{R}
}
\description{
Function to rotate X to match Y via singular value decomposition
}
