% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPIgen.R
\docType{data}
\name{PPIgen}
\alias{PPIgen}
\title{PPI genetic interactions}
\format{Binary adjacency matrix}
\usage{
PPIgen
}
\description{
The dataset contains a network formed by genetic protein-protein
interactions (PPI) between 67 Saccharomyces cerevisiae proteins.
The network is formed of 294 links.
The data were downloaded from the Biological General Repository for
Interaction Datasets (BioGRID) database \url{http://thebiogrid.org/}
}
\details{
Binary adjacency matrix containing genetic interactions between 67 proteins.
}
\references{
Gollini, I., and Murphy, T. B. (2016), 'Joint Modelling of Multiple Network Views', Journal of Computational and Graphical Statistics, 25(1), 246-265 \url{http://arxiv.org/abs/1301.3759}.
}
\seealso{
\code{\link{PPIphy}}
}
\keyword{datasets}
